# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para sugestões de presentes personalizadas.",
        "title": "Gift App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/reminders/{reminderId}": {
            "post": {
                "description": "Exemplo de payload: {\"userID\":\"a3f24e53-0d56-46d9-8ea2-0dbb5f64da8a\",\"friendID\":\"9b02ce54-4f42-4a8b-a539-5b53a6e37e63\",\"type\":\"anniversary\",\"triggerAt\":\"2026-09-20\",\"message\":\"Enviar flores\"}. Campo triggerAt no formato YYYY-MM-DD.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reminders"
                ],
                "summary": "Atualizar lembrete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do lembrete",
                        "name": "reminderId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados a atualizar",
                        "name": "reminder",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.ReminderUpsertRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Reminder"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{userId}/reminders": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reminders"
                ],
                "summary": "Listar lembretes do usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do usuário",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Reminder"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Exemplo de payload: {\"friendID\":\"9b02ce54-4f42-4a8b-a539-5b53a6e37e63\",\"type\":\"birthday\",\"triggerAt\":\"2026-08-15\",\"message\":\"Comprar presente ate uma semana antes\"}. Campo triggerAt no formato YYYY-MM-DD.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reminders"
                ],
                "summary": "Criar lembrete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do usuário",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do lembrete",
                        "name": "reminder",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.ReminderUpsertRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Reminder"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.Reminder": {
            "type": "object",
            "properties": {
                "friendID": {
                    "type": "string",
                    "example": "9b02ce54-4f42-4a8b-a539-5b53a6e37e63"
                },
                "message": {
                    "type": "string",
                    "example": "Comprar presente ate uma semana antes"
                },
                "reminderID": {
                    "type": "string",
                    "example": "d8c8efdf-c52f-4d6b-8e2e-b83f78de4f77"
                },
                "triggerAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-08-15T00:00:00Z"
                },
                "type": {
                    "type": "string",
                    "example": "birthday"
                },
                "userID": {
                    "type": "string",
                    "example": "5581a365-394f-467d-ae13-0d01e4cf1863"
                }
            }
        },
...
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "birthDate must be in format YYYY-MM-DD"
                }
            }
        },
...
        "http.ReminderUpsertRequest": {
            "type": "object",
            "properties": {
                "friendID": {
                    "type": "string",
                    "example": "9b02ce54-4f42-4a8b-a539-5b53a6e37e63"
                },
                "message": {
                    "type": "string",
                    "example": "Comprar presente ate uma semana antes"
                },
                "triggerAt": {
                    "type": "string",
                    "format": "date",
                    "example": "2026-08-15"
                },
                "type": {
                    "type": "string",
                    "example": "birthday"
                },
                "userID": {
                    "type": "string",
                    "example": "a3f24e53-0d56-46d9-8ea2-0dbb5f64da8a"
                }
            }
        },
...
    }
}